tests frame query generator and moved to lib
